backend/requirements.txt
alembic==1.5.2
boto3==1.16.63
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
ecdsa==0.14.1
fastapi==0.63.0
fire==0.4.0
h11==0.12.0
Mako==1.1.4
MarkupSafe==1.1.1
mypy==0.800
mypy-extensions==0.4.3
Pillow==8.2.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
requests==2.25.1
rsa==4.7
six==1.15.0
SQLAlchemy==1.3.22
sqlalchemy-stubs==0.4
starlette==0.13.6
typed-ast==1.4.2
typing-extensions==3.7.4.3
uvicorn==0.13.3
